pyproject.toml
[project]
name = "foundry-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "foundry-platform-sdk>=0.112.0",
    "mcp>=1.6.0",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_foundry"]

[project.scripts]
mcp-server-foundry = "mcp_server_foundry:main"
